rest api 리팩터링
