+++ src/main/java/com/shlee7131/financedemo/WebConfig.java
@@ -14,7 +14,6 @@ public class WebConfig implements WebMvcConfigurer {
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(authInterceptor)
                 .addPathPatterns("/**")
-                .excludePathPatterns("/login")
-                .excludePathPatterns("/register");
+                .excludePathPatterns("/auth/**");
     }
 }
