+++ src/main/java/com/shlee7131/financedemo/controller/AuthController.java
@@ -2,8 +2,11 @@ package com.shlee7131.financedemo.controller;
 
 import com.shlee7131.financedemo.exception.BadRequestException;
 import com.shlee7131.financedemo.service.AuthService;
+import com.shlee7131.financedemo.service.UserService;
 import com.shlee7131.financedemo.service.dto.UserReqDto;
+import com.shlee7131.financedemo.service.dto.UserRespDto;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -20,12 +23,14 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 public class AuthController {
     private final AuthService authService;
+    private final UserService userService;
 
-    @PostMapping("/login")
+    @PostMapping("/auth/local")
     public ResponseEntity<?> login(@Valid @RequestBody UserReqDto userReqDto, HttpServletResponse response) {
         Optional<String> sessionId = authService.login(userReqDto);
 
@@ -38,5 +43,16 @@ public class AuthController {
         response.addCookie(cookie);
 
         return new ResponseEntity<>(null, HttpStatus.OK);
+    }
+
+    @PostMapping("/auth/local/new")
+    public ResponseEntity<UserRespDto> register(@Valid @RequestBody UserReqDto userReqDto) {
+        Optional<UserRespDto> userRespDto = userService.createUser(userReqDto);
+        log.info("userReqDto: {}",userReqDto.toString());
+        log.info("userRespDto: {}", userRespDto.toString());
+
+        if (userRespDto.isEmpty()) throw new BadRequestException("이메일이 중복됩니다");
+
+        return new ResponseEntity<>(userRespDto.get(), HttpStatus.CREATED);
     }
 }
